Move rtps message to separate folder
