+++ dds/src/rtps/mod.rs
@@ -3,7 +3,6 @@ pub mod cache_change;
 pub mod error;
 pub mod message_receiver;
 pub mod message_sender;
-pub mod messages;
 pub mod reader_locator;
 pub mod reader_proxy;
 pub mod stateful_reader;
